tiling function makes tiles
names are in the styel of (current image in x dimension) x and (current
image in y dimension) y

May change convention later, but for now worked to show test.

Would be good to make a generalizable version of
get_tile_start_end_index so that we can greatly reduce the code side.

Could also use a condensed write tile function