offset multi-d, thresh based on image int
offset is now 2d, whereas before it was a single value for both x and y.
Slightly changes the offest.

Threshold of image intensity is now based on the max intensity of the
array, rather than the flat 255 value.  This may have been causing
problems and allowing too many ROIs in that would otherwise have been
thresholded out.