read, recast, linear regression
Task script with two functions and a partial working funciton.

recast_max_diff_90_deg casts one column to be always within 90 degrees
of the other columm

regress performs a linear regression on two columns, with x as the
second column and y as the first

Finally, analyze data currently generates a pandas dataframe using those
functions to get linear regression results for three modality pairs.

It also reads in the 'Cleaned data.csv' file, but that may be altered as
I change the conventions for the results files