fixed indexing and x y switch
Indexing starts at 0 for the roi and tile locs, so no need to do the -1.

Also fixed a bug introduced due to numpy array weirdness where it flips
order of variables - at least for sitk.  This flip only really 'matters'
for looking at images or any particular ROI, but fixing it would require
switching the (extremely large) set of data from X <-> Y.  Instead, for
just the purpose of images, we are doing the switch when generating the
image array and flipping it to the correct way when saving in sitk.