Delete initial origin change overlay
In the current script, overlays block the rest of the script from
running.  Due to this, the initial origin change overlay becomes
a problem when you don't wish the change the origin.  In addition, it is
unnecessary because their is a prior overlay in the rotation query.

The prior overlay is the same as the initial origin change overlay,
making it redundant.

This may become a problem in the future if the rotation/origin change
are decoupled.  However, for now, this saves some time and computing
resources.