Propogate type of transform from top level
The type_of_transform keyword lets you manipulate which transform is
used in the registration script, e.g. affine or euler.  However, it
was previously only callable from the 'register' and 'plot overlay'
functions, and wasn't referenced from higher level functions.

This meant that in practice, you had to change the default value of the
keyword to change the transform.

This commit propogates the keyword through all the registration
functions, allowing you to change it at the command line.