Define optimizer/metric/interpolator from top
Made a new function that creates a base image registration class,
modifying the optimizer, metric, and interpolator for use in a
registration task.  This new method allows us to define the variables
from the top of the script, though many keywords/arguments need to be
implemented to make this more flexible.