Remove sqrt from overlay image
This change removes the sqrt function that was used to window/level
the SHG/Polarimetry images.

This was done because the raw Mueller images have too close intensity,
and the sqrt makes it more difficult to see the differences.

Ideally, there would be a window/level command in place of the sqrt.
The todo addresses that.