Make func that copies necessary metadata
For now, it only changes the units by default.

Warning: the necessary keys variable is mutable and can change with
function calls, so there is also a todo to fix this using the *
capability in python.  However, I don't know it well enough to work
with it for this type of funciton call yet.