Get transform type str and associated tests
Writing a SimpleITK transform to a file and reading it back in makes
the type of the resulting transform into the generic Transform class,
instead of the actual transform type.

However, the underlying C++ class is still correct.

This function extracts the itk:: C++ class from the object str and
returns it.

Accompanying tests test it for Affine2D, Euler2D, and Bspline2D