Fix type check in change_transform_rotation
The type check was in the previous, nonworking form, where it tried
to directly check the type of the SimpleITK transform.  This doesn't
work because the type is the generic Transfrom after saving/reading
from file - you need to use the get_transform_type_str method to
get out the type info.