Fix change_transform b/c generic transform issue
The generic transform issue is where SimpleITK reads in a printed
transform as a generic Transform object instead of whichever actual
transform that was saved.

This temporary changes makes it so that we can change the rotation
for a generic transform using set/get parameters.

This will be reverted once an intermediary function is implemented
to set the transform type appropriately upon read in.