Move get image size to metadata and fix
The get image size function wasn't working for large image files,
and incurred a lot of overhead due to fully reading in the image.

This change moves it to the itk module, and implements the itk
file reader to get the image size.  The file reader can read in the size
without opening the actual image, so this has a lot less overhead
associated with it.