Implement func makes registration parameters dict
The number of possible keywords for registration parameters can
grow quite extensively.  Passing it through many different functions
with every keyword argument gets burdensome and confusing fast.

It is easier to predefine a dictionary, where the user sets the
parameters, and then pass it through as a single argument instead.

This function defines the dictionary, however, no other changes have
been made to make it useful.