Remove the imagej.py requirements for easy dist
The imagej.py code is not used for anything in the current project,
except for a single example script.

These packages complicate the installation considerably, and might
make it very difficult for people to download/use this project.

The requirements are being deleted until there is code that actually
depends on it, though they'll stay in the travis server.