Begin US image class in reconstruction.py
This class will handle all the input/processing internally, and will
make it an easier api than all the regular functions made outside of it.
In addition, it will be implemented alongside tests to make sure it is
working correctly.

Right now it only sets a few path variables, and has a tested function
to open the position list file.