Separate off Image funcs from assembler
This change is starting to make the US image assembler into just an
assembler, which will be able to work off of any kind of data.  E.g,
IQ data, RF data, B mode, or specially formatted data that requires
unique processing algorithms to work.

This will carry out and make it easier to get an image back overall.