Create OCT processing code
These changes enable OCT image registration and processing.  For example, the read mat code is now capable of reading V7.3 Matlab .m files, which are saved in an HDF format.

In addition, there is a task script for taking processed OCT tiles and turning them into 3D tifs.  Finally, there is a function to open the OCT image.